Tweak DrivePacket to set your own timeout
This means we can set short timeouts for things we intend to stop ASAP, such as
one second, but means we can set longer timeouts for actions we want to time.

My rationale for doing this was so that the API was more similar to the old
EnqueueMotionPacket, thus allowing us to use the old HolonomicRobotController
with minimal modification.